backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
anthropic==0.49.0
pdfplumber==0.11.5
python-docx==1.1.2
slowapi==0.1.9
python-dotenv==1.1.0
